+++ crates/stackable-shared/src/lib.rs
@@ -1,7 +1,8 @@
 //! This crate contains various shared helpers and utilities used across other crates in this
 //! workspace.
 
-pub mod commons;
 pub mod crd;
+pub mod secret;
+#[cfg(feature = "time")]
 pub mod time;
 pub mod yaml;