+++ crates/stackable-operator/CHANGELOG.md
@@ -21,7 +21,7 @@ All notable changes to this project will be documented in this file.
 - BREAKING: Some modules have been moved into the `stackable-shared` crate, so that they can also be
   used in `stackable-certs` and `stackable-webhook` ([#1074]):
   - The module `stackable_operator::time` has moved to `stackable_operator::shared::time`
-  - The module `stackable_operator::commons::secret` has moved to `stackable_operator::shared::commons::secret`
+  - The module `stackable_operator::commons::secret` has moved to `stackable_operator::shared::secret`
 
 ### Fixed